requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
PyJWT
bcrypt
python-dotenv
gunicorn
psycopg2-binary
